# Enable reading gzip files faster with the python main thread using python-isal.