--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -54,7 +54,7 @@ jobs:
       run: python -m pip install tox
     - name: Test
       run: tox -e py
-      if: !matrix.with-isal
+      if: matrix.with-isal == null
     - name: Test with isal
       run: tox -e isal
       if: matrix.with-isal