--- a/tox.ini
+++ b/tox.ini
@@ -12,6 +12,12 @@ commands =
     coverage report
     coverage xml
 
+[testenv:isal]
+deps =
+    pytest
+    coverage
+    isal
+
 [testenv:flake8]
 basepython = python3.6
 deps = flake8