--- a/tox.ini
+++ b/tox.ini
@@ -22,16 +22,19 @@ deps =
 basepython = python3.6
 deps = flake8
 commands = flake8 src/ tests/
+skip_install = true
 
 [testenv:mypy]
 basepython = python3.6
 deps = mypy
 commands = mypy src/
+skip_install = true
 
 [flake8]
 max-line-length = 99
 max-complexity = 10
 extend_ignore = E731
+skip_install = true
 
 [coverage:report]
 exclude_lines =