--- a/tox.ini
+++ b/tox.ini
@@ -27,7 +27,7 @@ skip_install = true
 [testenv:mypy]
 basepython = python3.6
 deps = mypy
-commands = mypy src/
+commands = mypy --ignore-missing-imports src/
 skip_install = true
 
 [flake8]