--- a/tox.ini
+++ b/tox.ini
@@ -27,14 +27,13 @@ skip_install = true
 [testenv:mypy]
 basepython = python3.6
 deps = mypy
-commands = mypy --ignore-missing-imports src/
+commands = mypy src/
 skip_install = true
 
 [flake8]
 max-line-length = 99
 max-complexity = 10
 extend_ignore = E731
-skip_install = true
 
 [coverage:report]
 exclude_lines =