This adds python-isal to the mix as an optional dependency. Optional, because it is only available on linux and requires extra libraries to be present.