@rhpvorderman 2020-12-16
Some benchmarks involving cutadapt in single-threaded mode:

Before:
```
$ hyperfine -w 3 -r  10 'cutadapt -a AGATCGGAAGAG 500000reads.fastq.qz -Z -o ramdisk/big2.cut.fastq.gz'
Benchmark #1: cutadapt -a AGATCGGAAGAG 500000reads.fastq.qz -Z -o ramdisk/big2.cut.fastq.gz
  Time (mean ± σ):      5.824 s ±  0.030 s    [User: 5.793 s, System: 0.029 s]
  Range (min … max):    5.785 s …  5.875 s    10 runs

$ du -h ramdisk/big2.cut.fastq.gz 
66M	ramdisk/big2.cut.fastq.gz

$ zcat ramdisk/big2.cut.fastq.gz | md5sum
97fc2a9c78e6597912a02b168e4cd068  -
```

After:
```
$ hyperfine -w 3 -r  10 'cutadapt -a AGATCGGAAGAG 500000reads.fastq.qz -Z -o ramdisk/big2.cut.fastq.gz'
Benchmark #1: cutadapt -a AGATCGGAAGAG 500000reads.fastq.qz -Z -o ramdisk/big2.cut.fastq.gz
  Time (mean ± σ):      4.010 s ±  0.067 s    [User: 3.973 s, System: 0.036 s]
  Range (min … max):    3.946 s …  4.190 s    10 runs

$ du -h ramdisk/big2.cut.fastq.gz 
62M	ramdisk/big2.cut.fastq.gz

$ zcat ramdisk/big2.cut.fastq.gz | md5sum
97fc2a9c78e6597912a02b168e4cd068  -
```
From 5.8 -> 4 seconds on average is quite a substantial improvement. Also the filesize is slightly smaller when using python-isal 66 -> 62m. So it is an approvement all around. Last but not least, the checksum of the content is the same, which is required.

This will be very very nice for people using cutadapt in single threaded mode. Galaxy users come to mind @bernt-matthias will probably like this.

If we get this into the conda release we can set the python-isal requirement for linux installs only. This will then be automatically available for biocontainers users as well on the next cutadapt release.