@marcelm 2020-12-18
Wow, the results looks really good!

This should definitely go in, just let me find some time to review the code changes, which may take until after the holidays. I hope that’s fine.