@rhpvorderman 2021-01-06
> I also was wondering about deployment. When this is merged, is it correct that the Conda package should be updated to depend on python-isal (on Linux)?

Yes, since that would propagate to the bioconda package for cutadapt as well. And as a result also the biocontainer. This will enable fast decompression/compression for a lot of users.

> Also, for the pip install case, have you considered providing wheels for python-isal? Then one would no longer need to install libisal-dev.

I have looked into this because of your comment. Initially I did not want to do it because it is a lot of hassle, and conda exists. I came to the conclusion that providing wheels adds no extra value. The isa-l libraries are still needed on the user system. So it does not eliminate a step for the user. I did however make the installation instructions more clear on the [python-isal readme](https://github.com/pycompression/python-isal) so it should be easier for users to install.

I changed the readme text on xopen to refer to this updated installation text.

> Or you could go even one step further by linking statically against libisal (thus including it in the wheel), and then the user doesn’t need to use their distribution package manager at all.

That would be a good solution. The dynamic linking works very well with conda and gives the benefit of automatically updating any bugs in isa-l away.
Static linking would however be more practical for more python users and would lead to less compatibility nightmares for me as the required version is simply included. The pace of change in isa-l is not that big and therefore it is very easy to keep up. There are however license issues that need to be discussed with the project maintainers before I am able to redistribute isa-l in that form. So I will do that.