@marcelm 2021-01-07
I agree: If you provide wheels at all, they should include a statically linked isa-l.

Another reason: Many people still use `pip` to get Cutadapt (judging from the numbers for last month at https://pypistats.org/packages/cutadapt vs https://anaconda.org/bioconda/cutadapt/files), and almost all of these will not get the improvements automatically because `python-isal` is just an optional dependency. If there was a wheel, it could be made a hard dependency instead (using a suitable environment marker so you only require it on Linux).